+++ cognite/neat/rules/importer/_base.py
@@ -34,7 +34,7 @@ class BaseImporter(ABC):
         self,
         return_report: bool = False,
         skip_validation: bool = False,
-        validators_to_skip: list[str] | None = None,
+        validators_to_skip: set[str] | None = None,
     ) -> tuple[Rules | None, list[ErrorDetails] | None, list | None] | Rules:
         """
         Creates `Rules` object from the data.